+++ database/drivers/godror/driver.go
@@ -1,0 +1,4 @@
+// Package godror registers the Oracle godror database/sql driver for Way applications.
+package godror
+
+import _ "github.com/godror/godror"