+++ database/drivers/mysql/driver.go
@@ -1,0 +1,4 @@
+// Package mysql registers the MySQL database/sql driver for Way applications.
+package mysql
+
+import _ "github.com/go-sql-driver/mysql"