+++ database/drivers/pgx/driver.go
@@ -1,0 +1,4 @@
+// Package pgx registers the pgx database/sql driver for Way applications.
+package pgx
+
+import _ "github.com/jackc/pgx/v5/stdlib"