+++ database/drivers/sqlite/driver.go
@@ -1,0 +1,4 @@
+// Package sqlite registers the SQLite database/sql driver for Way applications.
+package sqlite
+
+import _ "github.com/mattn/go-sqlite3"