+++ database/drivers/sqlserver/driver.go
@@ -1,0 +1,4 @@
+// Package sqlserver registers the Microsoft SQL Server database/sql driver for Way applications.
+package sqlserver
+
+import _ "github.com/denisenkom/go-mssqldb"